remove new producer and replace old one with same code
