--- code/lambdas/fragmentation_producer/requirements.txt
@@ -1,5 +1,0 @@
-boto3
-kafka-python
-pandas
-faker
-aws-msk-iam-sasl-signer-python
